Split SearchForm into own component
