+++ src/components/Hero.css
@@ -3,13 +3,28 @@
     justify-content: center;
     align-items: center;
     width: 100%;
-    max-width: 450px;
-    margin: auto;
-    /* background-image: url('../assets/images/library.png'); */
-    background-size: 100% 100%;
-    background-position: center center;
+    background-image: url('../assets/images/library.png');
+    background-size: center;
+    background-position: center;
+    padding-top: 20px;
+    padding-bottom: 30px;
   }
   
-  .hero-form .btn {
+  .hero-container {
+    width: 100%;
+    max-width: 450px;
+  }
+  
+  .hero-text {
+    max-width: 450px;
+    margin: auto;
+    color: rgb(255, 255, 255);
+  }
+  
+  .search-form {
+    max-width: 450px;
+    margin: auto;
+  }
+  .search-form .btn {
     margin-top: 10px;
   }
