+++ src/components/Hero.js
@@ -1,44 +1,27 @@
 import './Hero.css';
-import React, { useState } from 'react';
-import { Container, Row, Col, Form, Button } from 'react-bootstrap';
+import React from 'react';
+import { Container, Row, Col } from 'react-bootstrap';
+import SearchForm from './SearchForm';
 
 const Hero = () => {
-  const [searchTerm, setSearchTerm] = useState('');
-
-  const handleSearch = (event) => {
-    event.preventDefault();
-    // Perform search using the `searchTerm` state variable
+  const handleSearch = (searchTerm) => {
     console.log(`Searching for "${searchTerm}"`);
-  };
-
-  const handleInputChange = (event) => {
-    setSearchTerm(event.target.value);
   };
 
   return (
     <div className="hero">
-      <Container>
-        <Row>
-          <Col>
+      <Container fluid>
+        <Row className="justify-content-center">
+          <Col xs={12} md={8} lg={7}>
             <div className="hero-text">
               <h1 className="hero-title">Whats the Word?</h1>
               <p className="hero-subtitle">Search for a word, any word, make it fun? make it serious? We've got you covered!</p>
             </div>
           </Col>
         </Row>
-        <Row>
-          <Col>
-            <Form className="hero-form" onSubmit={handleSearch}>
-              <Form.Control
-                type="text"
-                placeholder="Enter a word"
-                value={searchTerm}
-                onChange={handleInputChange}
-              />
-              <Button variant="primary" type="submit">
-                Search
-              </Button>
-            </Form>
+        <Row className="justify-content-center">
+          <Col xs={12} md={8} lg={7}>
+            <SearchForm onSearch={handleSearch} />
           </Col>
         </Row>
       </Container>
